requirements.txt
mcp
requests
gspread
python-telegram-bot
python-dotenv
